requirements.txt
accelerate==1.8.1
antlr4-python3-runtime==4.9.3
black==24.4.2
# certifi==2025.6.15
# charset-normalizer==3.4.2
colorama==0.4.6
dataset #==2.19.1
dotenv==0.9.9
evaluate==0.4.2
filelock==3.18.0
# fsspec==2025.5.1
hf-xet==1.1.5
huggingface-hub==0.33.2
hydra-core==1.3.2
# idna==3.10
ipykernel==6.29.4
ipython==8.24.0
isort==5.13.2
# jinja2==3.1.6
# jupyter==1.0.0
markupsafe==3.0.2
# matplotlib==3.8.4
mpmath==1.3.0
networkx==3.4.2
nltk==3.8.1
numpy==1.26.4
omegaconf==2.3.0
# optimum-quanto==0.2.7
packaging==25.0
psutil==7.0.0
pyarrow-hotfix==0.7
python-dotenv==1.0.1
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.4
ruff==0.4.3
safetensors==0.5.3
sympy==1.14.0
tensordict==0.8.3
tokenizers==0.21.2
# -e file:///home/ucabpag/Hercules/fake_torch_package
tqdm==4.67.1
transformers==4.53.0
typing-extensions==4.14.0
urllib3==2.5.0
tyro
